requirements.txt
streamlit
pycaret[full]
streamlit-pandas-profiling
pandas-profiling